+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Added
+- extend `TypeContext` creation to support configuration of differing annotation inheritance by annotation type
+
+### `jsonschema-module-jakarta-validation`
+#### Fixed
+- consider inheritance of validation constraint annotations
+- when limiting scope to validation groups, also consider a specified group's supertypes 
+
+### `jsonschema-module-javax-validation`
+#### Fixed
+- consider inheritance of validation constraint annotations
+- when limiting scope to validation groups, also consider a specified group's supertypes
 
 ## [4.30.0] - 2023-04-16
 ### `jsonschema-generator`